:pencil: Renaming tests
