+++ omisego-admin/src/liveTest/kotlin/co/omisego/omisego/features/ListTokenTest.kt
@@ -9,8 +9,9 @@ package co.omisego.omisego.features
 
 import co.omisego.omisego.LiveTest
 import co.omisego.omisego.model.params.LoginParams
-import co.omisego.omisego.model.transaction.list.TransactionListParams
+import co.omisego.omisego.model.params.TokenListParams
 import org.amshove.kluent.shouldBe
+import org.amshove.kluent.shouldBeGreaterThan
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -19,7 +20,7 @@ import org.robolectric.annotation.Config
 
 @RunWith(RobolectricTestRunner::class)
 @Config(sdk = [23])
-class ListTransaction : LiveTest() {
+class ListTokenTest : LiveTest() {
     private val secret by lazy { loadSecretFile("secret.json") }
 
     @Before
@@ -34,9 +35,9 @@ class ListTransaction : LiveTest() {
     }
 
     @Test
-    fun `list transaction should be returned successfully`() {
-        val response = client.getTransactions(
-            TransactionListParams.create(
+    fun `list tokens should be returned successfully`() {
+        val response = client.getTokens(
+            TokenListParams.create(
                 searchTerm = null
             )
         ).execute()
@@ -44,20 +45,6 @@ class ListTransaction : LiveTest() {
         response.isSuccessful shouldBe true
         response.body()?.data?.pagination?.perPage shouldBe 10
         response.body()?.data?.pagination?.currentPage shouldBe 1
-        response.body()?.data?.data?.size shouldBe 10
-    }
-
-    @Test
-    fun `list transaction with a specific account should return transactions associated with the account`() {
-        val response = client.getTransactions(
-            TransactionListParams.create(
-                searchTerm = secret.getString("account_address")
-            )
-        ).execute()
-
-        response.isSuccessful shouldBe true
-        response.body()?.data?.data?.forEach {
-            (secret.getString("account_address") in arrayOf(it.from.address, it.to.address)) shouldBe true
-        }
+        response.body()?.data?.data?.size!! shouldBeGreaterThan 0
     }
 }
