+++ omisego-admin/src/liveTest/kotlin/co/omisego/omisego/features/ListTransactionTest.kt
@@ -9,9 +9,8 @@ package co.omisego.omisego.features
 
 import co.omisego.omisego.LiveTest
 import co.omisego.omisego.model.params.LoginParams
-import co.omisego.omisego.model.params.WalletParams
+import co.omisego.omisego.model.transaction.list.TransactionListParams
 import org.amshove.kluent.shouldBe
-import org.amshove.kluent.shouldNotBe
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -20,7 +19,7 @@ import org.robolectric.annotation.Config
 
 @RunWith(RobolectricTestRunner::class)
 @Config(sdk = [23])
-class GetWallet : LiveTest() {
+class ListTransactionTest : LiveTest() {
     private val secret by lazy { loadSecretFile("secret.json") }
 
     @Before
@@ -35,24 +34,30 @@ class GetWallet : LiveTest() {
     }
 
     @Test
-    fun `get a user wallet should be returned associated user of the account`() {
-        val response = client.getWallet(
-            WalletParams(secret.getString("user_address"))
+    fun `list transaction should be returned successfully`() {
+        val response = client.getTransactions(
+            TransactionListParams.create(
+                searchTerm = null
+            )
         ).execute()
 
         response.isSuccessful shouldBe true
-        response.body()?.data?.user shouldNotBe null
-        response.body()?.data?.account shouldBe null
+        response.body()?.data?.pagination?.perPage shouldBe 10
+        response.body()?.data?.pagination?.currentPage shouldBe 1
+        response.body()?.data?.data?.size shouldBe 10
     }
 
     @Test
-    fun `get an account wallet should be returned associated account of the account`() {
-        val response = client.getWallet(
-            WalletParams(secret.getString("account_address"))
+    fun `list transaction with a specific account should return transactions associated with the account`() {
+        val response = client.getTransactions(
+            TransactionListParams.create(
+                searchTerm = secret.getString("account_address")
+            )
         ).execute()
 
         response.isSuccessful shouldBe true
-        response.body()?.data?.account shouldNotBe null
-        response.body()?.data?.user shouldBe null
+        response.body()?.data?.data?.forEach {
+            (secret.getString("account_address") in arrayOf(it.from.address, it.to.address)) shouldBe true
+        }
     }
 }
